use hashing function exposed by browserify
As of version 3.31.5 browserify uses relative paths to compute the file
hash. This means the karma-browserify bundle is incorrectly
hashing file ids to be required by the tests.

The hashing function is now available directly on the browserify bundle
instance (see substack/node-browserify@9499ffb9). This commit fixes the
above mentioned issue by delegating the file id hashing to this exposed
function.

At the moment though, that hashing function is prefixed with an `_`
possibly indicating that it's intended for private internal usage. So
any changes to this methods name or placement will break this code. In
that case though, the break will be much easier to detect than the
errors caused previously.